requirements.txt
asyncio~=3.4.3
crawl4ai~=0.4.248
exa-py
newsdataapi~=0.1.21
newspaper3k
python-dotenv~=1.0.1
phidata~=2.7.10
regex~=2024.11.6
tweepy~=4.15.0
pycoingecko~=3.2.0
duckduckgo-search
pandas~=2.2.2
typing~=3.7.4.3
pydantic~=2.10.4
agno~=1.0.8
langchain-core
requests~=2.32.3
langchain-openai~=0.3.3
langchain~=0.3.13
chromadb
lxml_html_clean
sentence_transformers
langchain_huggingface
ftfy~=6.3.1
langchain-community
nest-asyncio~=1.6.0
faiss-cpu
openai~=1.63.2
googlesearch-python
pycountry
langgraph
fal_client
